+++ src/Sienar.Architecture.Web/Ui/SienarApp.razor
@@ -3,8 +3,6 @@
 @using static Microsoft.AspNetCore.Components.Web.RenderMode
 @using Sienar.Infrastructure
 
-@namespace Sienar.Components
-
 @inject IStyleProvider StyleProvider
 @inject IScriptProvider ScriptProvider
