+++ src/Sienar.Plugin.Cms.Client/Identity/Processors/LoadUserDataProcessor.cs
@@ -1,11 +1,11 @@
 ﻿#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 
 using System.Threading.Tasks;
-using Sienar.Components;
 using Sienar.Data;
 using Sienar.Hooks;
 using Sienar.Infrastructure;
 using Sienar.Processors;
+using Sienar.Ui;
 
 namespace Sienar.Identity.Processors;
