+++ src/Sienar.Plugin.Cms.Client/Infrastructure/InitializeCsrfTokenOnAppStartHook.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
-using Sienar.Components;
 using Sienar.Hooks;
+using Sienar.Ui;
 
 namespace Sienar.Infrastructure;
