+++ src/Sienar.Utils/Sienar.Utils.csproj
@@ -35,10 +35,4 @@
 		<PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="8.0.0"/>
 	</ItemGroup>
 
-	<ItemGroup>
-	  <AdditionalFiles Include="Components\SienarHead.razor" />
-	  <AdditionalFiles Include="Components\SienarRoutes.razor" />
-	  <AdditionalFiles Include="Components\_Imports.razor" />
-	</ItemGroup>
-
 </Project>
